Duplicate Timestamps
FEWSNET suffered from accidental duplicate timestamps. This modified version isolates only those timestamps which fall outside of the expected 00, 15, 30, and 45 minute intervals. Although a duplicate timestamp could fall on one of those 15-minute intervals (e.g. 59, 00 or 15, 16), for simplicity I will be keeping them in for now. It is yet to be determined whether that edge case produces valid timestamps on the 15-minute interval if it is the result of a duplicate timestamp.